perf(sessions): isolate archive Worker imports
Keep read-only archive materialization and publication from loading session reclamation and plugin ownership code. Reuse the existing file helpers from a leaf module and load reclamation only for the reclaim operation.

Preserve the guarded commit authorization and settlement protocol introduced by #137672 by moving its complete current Worker implementation into the lazy runtime. Keep the parent queue, commit-request handling, error precedence, generation fences and cleanup outcomes unchanged.

The original controlled Linux pair reduced warm deletion median from 895ms to 398ms on d84. That timing is historical; this integration does not claim a new latency measurement.
